Move another function to `azure-pipelines.js`
This is another cleanup patch.

This patch is best viewed using `--color-moved`.
